Support DeHackEd patches with BOOM extensions (BEX)
Completely removed first attempt at supporting DeHackEd patches in
commit 0354dc3f12511c9f9e8e77ad7ce35a56bf689db8, and instead of using
code from Chocolate DOOM, I've taken code from WinMBF, allowing support
for BEX as well. A lot of work is still to be done, but as an initial
"test", map names in BTSX_E1.WAD now appear in the automap.

+++ src/info.h
@@ -1159,12 +1159,12 @@ typedef enum
 typedef struct
 {
     spritenum_t sprite;
-    int         frame;
-    int         tics;
+    long        frame;
+    long        tics;
     actionf_t   action;
     statenum_t  nextstate;
-    int         misc1;
-    int         misc2;
+    long        misc1;
+    long        misc2;
 } state_t;
 
 typedef struct
